WPB-26096: Move quality metric calculation into stats
Move quality index calulation into stats object. This is to be
preperation for windowing operation to be implemented.
