+++ include/avs.h
@@ -42,7 +42,6 @@ extern "C" {
 #include "avs_conf_pos.h"
 #include "avs_conf_member.h"
 #include "avs_dict.h"
-#include "avs_stats.h"
 #include "avs_jzon.h"
 #include "avs_kase.h"
 #include "avs_log.h"
@@ -52,7 +51,8 @@ extern "C" {
 #include "avs_frame_encryptor.h"	
 #include "avs_frame_decryptor.h"	
 #include "avs_zapi.h"
-#include "avs_icall.h"	
+#include "avs_icall.h"
+#include "avs_stats.h"
 #include "avs_iflow.h"	
 #include "avs_peerflow.h"	
 #include "avs_jsflow.h"
