+++ src/peerflow/pc_platform_default.c
@@ -19,7 +19,6 @@
 #include <re.h>
 #include <avs_zapi.h>
 #include <avs_base.h>
-#include <avs_stats.h>
 #include <avs_icall.h>
 #include <avs_keystore.h>
 #include <avs_iflow.h>
